requirements.txt
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.3
django-filter==2.1.0
djangorestframework==3.9.4
drf-yasg==1.16.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
Markdown==3.1.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
pyrsistent==0.15.2
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
ruamel.yaml==0.15.97
six==1.12.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
uritemplate==3.0.0
urllib3==1.25.3
